More Octree and Region updates
Updated the region class so that it now defines a __getitem__, allowing it to replace all my sloppy bounds variables with the object instead. Also updated the Octree class to define the regions that we are looking at and separate the points into buckets.
TODO: Still need to add logic for when we reach a leaf node, buckets with just one point, so that it is not going forever into recursively smaller buckets of just one point.